Add files via upload
This project is to analyse the electric vehicles using the data set that is sourced from Kaggle. As, the world shifts towards cleaner and more sustainable mobility solutions by understanding this adapting the electric vehicles is becoming important. This project centres on leveraging data mining techniques to analyse the Electric Vehicle Population dataset, focusing on Washington State, to unearth patterns, trends, and factors affecting the growth and sustainability of EV usage.